New server Cvar: sv_extensions. When set to zero, it disables the extended challenge, causing QF clients to not detect a QuakeForge server. This makes the client not add the QF-specific info strings, in turn making it not overflow original QW clients' limited setinfo space. This also probably makes certain proxies work with QF servers again. It /also/ has the side effect of disabling compressed downloads. If you want compressed downloads, you need sv_extensions on.
Also, remove the "stdver" setinfo in the client -- it's handled by
*qsg_version if there's a QF server on the other end, so it was just
wasted space.

+++ include/client.h
@@ -285,7 +285,7 @@ typedef struct
 
 	char		levelname[40];	// for display on solo scoreboard
 	int			playernum;
-	int			stdver;
+	float		stdver; 		// QSG version
 
 // refresh related state
 	struct model_s	*worldmodel;	// cl_entitites[0].model
